Pass the local buffer cmd and args to subprocess
Any file local or dir local variables should be passed from the
current jsonnet buffer to the subprocess we are executing.  Since they
will all be buffer local, we can do this by moving the let capturing
the command and args to outside the `with-current-buffer` that is
operating on the output buffer `*jsonnet-output*`.

This allows us to evaluate a buffer like this,
```
{
  hello: std.extVar('foo'),
}

// Local Variables:
// eval: (setq-local jsonnet-command-options '("--ext-str" "foo=world"))
// End:
```